+++ src/main/java/com/wellnesswave/WellnessWave/Repository/PatientRepository.java
@@ -3,5 +3,5 @@ package com.wellnesswave.WellnessWave.Repository;
 import com.wellnesswave.WellnessWave.Entities.Patient;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PatientRepository extends JpaRepository<Patient, Long> {
+public interface PatientRepository extends JpaRepository<Patient, Integer> {
 }
